+++ packages/ts-components/src/components/olympics/schedule/OlympicsSchedule.tsx
@@ -1,18 +1,16 @@
-import React, { FC, useEffect } from 'react';
+import React, { FC, useEffect, useState } from 'react';
 
 import { Container } from './styles';
-import { colours } from '@times-components/styleguide';
+import { HeadingContainer, Label, Button, Heading } from '../shared-styles';
 
-import { OlympicsKeys } from '../types';
 import { injectScript } from '../../../helpers/widgets/inject-script';
+import { OlympicsKeys, config } from '../OlympicsKeys';
 
 export const OlympicsSchedule: FC<{
-  keys: OlympicsKeys;
-  sectionColor?: string;
+  keys?: OlympicsKeys;
   inArticle?: boolean;
 }> = ({
-  keys: { endpoint, authToken, gamesCode },
-  sectionColor = colours.section.sport,
+  keys: { endpoint, authToken, gamesCode } = config.prod,
   inArticle = true
 }) => {
   useEffect(() => {
@@ -36,13 +34,27 @@ export const OlympicsSchedule: FC<{
     );
   }, []);
 
+  const [showAll, setShowAll] = useState(false);
+  const handleShowAll = () => {
+    setShowAll(!showAll);
+  };
+
   return (
-    <Container sectionColour={sectionColor} inArticle={inArticle}>
+    <Container inArticle={inArticle} showAll={showAll}>
+      <HeadingContainer>
+        <Label>Tokyo 2020</Label>
+        <Heading>Event Schedule</Heading>
+      </HeadingContainer>
       <div
         className="pa-schedule"
         data-auth-token={authToken}
         data-games-code={gamesCode}
       />
+      <div className="buttonContainer">
+        <Button onClick={handleShowAll}>
+          {showAll ? 'Collapse' : 'Show All'}
+        </Button>
+      </div>
     </Container>
   );
 };