+++ packages/article-extras/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.27](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.7.26...@times-components/article-extras@0.7.27) (2021-07-28)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
+## [0.7.26](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.7.25...@times-components/article-extras@0.7.26) (2021-07-28)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
+## [0.7.25](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.7.24...@times-components/article-extras@0.7.25) (2021-07-28)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
+## [0.7.24](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.7.23...@times-components/article-extras@0.7.24) (2021-07-27)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
 ## [0.7.23](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.7.22...@times-components/article-extras@0.7.23) (2021-07-26)
 
 **Note:** Version bump only for package @times-components/article-extras